--- source/scheduler/_tests/_startScheduler.test.ts
@@ -1,22 +1,0 @@
-const mockSchedule = { scheduleJob: jest.fn() }
-jest.mock("node-schedule", () => mockSchedule)
-
-jest.mock("../../db/getDB")
-import { MockDB } from "../../db/__mocks__/getDB"
-import { getDB } from "../../db/getDB"
-const mockDB = getDB() as MockDB
-
-import { generateInstallation } from "../../testing/installationFactory"
-import { runSchedule } from "../startRepeatedTaskScheduler"
-
-it("runs scheduleJob for your tasks", async () => {
-  const scheduler = {
-    "1 2 3 4 5": "every_so_often.ts",
-  }
-
-  mockDB.getSchedulableInstallations.mockResolvedValueOnce([generateInstallation({ scheduler })])
-
-  await runSchedule()
-
-  expect(mockSchedule.scheduleJob).toBeCalledWith("1 2 3 4 5", expect.anything())
-})
